refactor: Create Frame package
