+++ src/main/java/bowling/domain/frame/FinalFrame.java
@@ -1,23 +1,26 @@
-package bowling.domain;
+package bowling.domain.frame;
 
+import bowling.domain.Score;
+import bowling.domain.ScoreState;
 import bowling.exception.CannotCalculateException;
 import bowling.exception.FrameTryException;
 
-public class NormalFrame extends Frame {
+public class FinalFrame extends Frame {
 
-    private static final int MAX_PIN_COUNT = 10;
-    private static final int MAX_TRY_COUNT = 2;
+    private static final int MAX_PIN_COUNT = 30;
+    private static final int MAX_TRY_COUNT = 3;
+    private static final int NOT_BONUS_MAX_COUNT = 10;
+    private static final int SECOND_TRY = 2;
     private static final String HIT_COUNT_EXCEPTION_MESSAGE = String.format("핀은 최대 %d개 까지 쓰러트릴 수 있습니다", MAX_PIN_COUNT);
-    private static final String TRY_COUNT_EXCEPTION_MESSAGE = String.format("최대 %d번 까지 시도할 수 있습니다", MAX_TRY_COUNT);
+    private static final String TRY_COUNT_EXCEPTION_MESSAGE = String.format("스트라이크 일 때 최대 %d번, 아닐 시 최대 %d번 까지 시도할 수 있습니다", MAX_TRY_COUNT, SECOND_TRY);
     private static final String CANNOT_CALCULATE_MESSAGE = "앞 투구가 끝나지 않아 계산 할 수 없습니다.";
 
-
-    private NormalFrame() {
+    private FinalFrame() {
         super();
     }
 
-    public static NormalFrame from() {
-        return new NormalFrame();
+    public static FinalFrame getFrame() {
+        return new FinalFrame();
     }
 
     @Override
@@ -29,8 +32,16 @@ public class NormalFrame extends Frame {
     }
 
     @Override
+    public void createScore() {
+        score = Score.of(pins.totalCount(), ScoreState.ofNone());
+    }
+
+    @Override
     public boolean roundEnded() {
-        return pins.tryCount() == MAX_TRY_COUNT || pins.isStrike();
+        if (pins.tryCount() == SECOND_TRY) {
+            return hasNoBonusTry();
+        }
+        return pins.tryCount() >= MAX_TRY_COUNT;
     }
 
     @Override
@@ -47,16 +58,7 @@ public class NormalFrame extends Frame {
         }
     }
 
-    @Override
-    public void createScore() {
-        if (pins.isStrike() && isFirstTry()) {
-            score = Score.of(MAX_PIN_COUNT, ScoreState.ofStrike());
-            return;
-        }
-        if (pins.totalCount() == MAX_PIN_COUNT) {
-            score = Score.of(MAX_PIN_COUNT, ScoreState.ofSpare());
-            return;
-        }
-        score = Score.of(pins.totalCount(), ScoreState.ofNone());
+    private boolean hasNoBonusTry() {
+        return pins.totalCount() < NOT_BONUS_MAX_COUNT;
     }
 }
