+++ src/test/java/bowling/domain/frame/FrameTest.java
@@ -1,5 +1,9 @@
-package bowling.domain;
+package bowling.domain.frame;
 
+import bowling.domain.frame.FinalFrame;
+import bowling.domain.frame.Frame;
+import bowling.domain.frame.Frames;
+import bowling.domain.frame.NormalFrame;
 import bowling.exception.CannotCalculateException;
 import bowling.exception.FrameTryException;
 import org.junit.jupiter.api.DisplayName;
